wheather/requirements.txt
Django==5.2.2
djangorestframework==3.16.0
python-decouple==3.8
Requests==2.32.3